+++ crates/vela_engine/src/tests/source_reload/dir_basic.rs
@@ -121,7 +121,7 @@ pub fn grant() {
     std::fs::write(
         game_dir.join("config.vela"),
         r#"
-pub const BONUS: int = 6;
+pub const BONUS: i64 = 6;
 "#,
     )
     .expect("write config module");
@@ -379,7 +379,7 @@ fn runtime_stages_dir_public_function_addition_until_safe_point() {
 #[test]
 fn runtime_stages_dir_return_abi_rejection_until_safe_point() {
     let root = unique_test_dir("runtime_stage_dir_return_abi");
-    let reward_file = write_typed_reward_modules(&root, "return grant();", "int", "2");
+    let reward_file = write_typed_reward_modules(&root, "return grant();", "i64", "2");
     let engine = Engine::builder()
         .execution_profile(ExecutionProfile::trusted())
         .build()
@@ -402,7 +402,7 @@ fn runtime_stages_dir_return_abi_rejection_until_safe_point() {
         Ok(OwnedValue::Scalar(vela_common::ScalarValue::I64(2)))
     );
 
-    write_typed_reward_module(&reward_file, "float", "6.0");
+    write_typed_reward_module(&reward_file, "f64", "6.0");
     runtime
         .stage_hot_reload_update_dir(&root)
         .expect("runtime should be hot-reload enabled")
@@ -441,8 +441,8 @@ fn runtime_stages_dir_return_abi_rejection_until_safe_point() {
         panic!("expected changed function return ABI");
     };
     assert_eq!(function, "game::reward::grant");
-    assert_eq!(old.as_deref(), Some("int"));
-    assert_eq!(new.as_deref(), Some("float"));
+    assert_eq!(old.as_deref(), Some("i64"));
+    assert_eq!(new.as_deref(), Some("f64"));
     assert!(source_span.is_some());
     assert_eq!(
         runtime.call_raw(
